feat: add ValidateUpdateShopCreatedStatus and UpdateShopCreatedStatus functions and add UpdateCreatedStatusShop struct

+++ helpers/static_data.go
@@ -1,6 +1,6 @@
 package helpers
 
-var CreatedStatuses map[string]int = map[string]int{
+var CreatedStatuses map[string]int8 = map[string]int8{
 	"wait":     0,
 	"rejected": 1,
 	"success":  2,
